requirements.txt
streamlit
pandas
selenium
webdriver-manager
plotly-express
apscheduler
requests-html

